+++ server/utils/network.go
@@ -1,11 +1,13 @@
 package utils
 
 import (
+	"context"
 	"encoding/json"
 	"os"
 	"os/exec"
 	"regexp"
 	"strings"
+	"time"
 )
 
 const ResolvConfBackup = "/etc/resolv.conf.bak"
@@ -71,7 +73,10 @@ func GetHostGateway() string {
 func GetIpInfo(ipInfo map[string]interface{}) error {
 	LogLn("get ip info")
 	// https://worldtimeapi.org/api/ip
-	cmd := exec.Command("/usr/bin/wget", "--no-check-certificate", "-q", "-O", "-", "https://ipinfo.io/json")
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+
+	cmd := exec.CommandContext(ctx, "/usr/bin/wget", "--no-check-certificate", "-q", "-O", "-", "https://ipinfo.io/json")
 	out, err := cmd.CombinedOutput()
 	if err != nil {
 		LogLn(string(out))